Fix broken 64 bit build

+++ std/internal/math/biguintnoasm.d
@@ -50,7 +50,7 @@ unittest
     uint [] a = new uint[40];
     uint [] b = new uint[40];
     uint [] c = new uint[40];
-    for (size_t i = 0; i < a.length; ++i)
+    for (uint i = 0; i < a.length; ++i)
     {
         if (i&1) a[i]=cast(uint)0x8000_0000 + i;
         else a[i]=cast(uint)i;
